refactor(wing): extract shared types to wing_types_shared.go (eliminate duplicate-and-sync risk)
